moved the training region out of the airSuspensionUtil and into it's own file
