+++ OASMan_ESP32/src/airSuspensionUtil.h
@@ -46,10 +46,5 @@ namespace PressureSensorCalibration
     void learnPressureSensorsRoutine();
 }
 
-// Refit any offset model whose sample count changed. Called at boot and periodically from task_trainAI.
-void trainOffsetModels();
-// Predicted true bag pressure from live flowing readings: rawBag + faded (default -> trained) offset.
-double getPredictedBagPressure(SOLENOID_AI_INDEX aiIndex, double raw_bag, double raw_tank);
-double getPredictedBagHeight(double raw_level);
 void setupLEDs();
 #endif
