+++ OASMan_ESP32/src/bluetooth/ble.cpp
@@ -509,8 +509,6 @@ uint8_t att_server_notify_SAFE(hci_con_handle_t con_handle, uint16_t attribute_h
     return att_server_notify(con_handle, attribute_handle, value, value_len);
 }
 
-extern uint8_t AIPercentage;
-
 ConfigValuesPacket buildCurrentConfigValuesPacket()
 {
     uint32_t configFlagsBits = 0;
